Separated out source retrieval from dashboard
